Check obstacles and add BASIC stub
Checks the obstacles and changes direction accordingly.
There is still some bug in it as it went haywire when it hit an obstacle right at the top border. Or so it seemed.
And bouncing off obstacles is also a bit weird.

As the program does not fit into the cassette buffer anymore I added a BASIC stub and placed it right after that into BASIC memory.